+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## History
-* 2021/02/12 ver2.25.2
+* 2021/02/** ver2.25.5
+	* Added 4x4 sudoku support (for kids in the school)
+	* Cross Helper feature added to Yajilin Composite Mode (Right click on Edges)
 	* Settings button, Dark Theme
+* 2021/02/21 ver2.25.4
+	* Added "Edge Free" and "Line Free" to tab settings.
+	* Made some improvements to "URL for contests".
+	* Load URL now accepts gmpuzzles link as well.
+* 2021/02/21 ver2.25.3
+	* Improved "Star battle" composite mode. One can mark dots between two/four cells.
+	* Enabled "Special" mode in Solving mode.
+	* Improved Sudoku Solving mode for French Keyboard Layout and possibly any keyboard layout.
+* 2021/02/16 ver2.25.2
+	* Added URL for competitive solving button.
+	* Save submode, style settings for various modes.
+	* Implemented Edge x composite mode.
+	* Fixed SHIFT + number bug (hopefully).
 * 2021/02/07 ver2.25.1
 	* Added Right Triangle, Left triangle and Darts to the Shapes. Merging the latest changes from upupstream.
 * 2021/01/19 ver2.24.22